UPSTREAM: LSM: Introduce enum lsm_order
In preparation for distinguishing the "capability" LSM from other LSMs, it
must be ordered first. This introduces LSM_ORDER_MUTABLE for the general
LSMs and LSM_ORDER_FIRST for capability. In the future LSM_ORDER_LAST
for could be added for anything that must run last (e.g. Landlock may
use this).

Change-Id: Ic09db63799ad4c680ce50b44312d32447ca5919f
Signed-off-by: Kees Cook <keescook@chromium.org>
Signed-off-by: Mr-Morat <mtarom124@gmail.com>