bluetooth: check key sizes only when Secure Simple Pairing is enabled
The encryption is only mandatory to be enforced when both sides are using
Secure Simple Pairing and this means the key size check makes only sense
in that case.

On legacy Bluetooth 2.0 and earlier devices like mice the encryption was
optional and thus causing an issue if the key size check is not bound to
using Secure Simple Pairing.

Signed-off-by: Mr-Morat <mtarom124@gmail.com>
Signed-off-by: manipvlator <mr.morat14@gmail.com>